wip, layout.tsx server component
