+++ dashboard/app/middleware.ts
@@ -1,0 +1,3 @@
+import { withMiddlewareAuthRequired } from "@auth0/nextjs-auth0/edge";
+
+export default withMiddlewareAuthRequired();
